apps/agent/requirements.txt
# LiveKit Agents SDK v1.3.7+ (requires Python ≤3.13)
# v1.3.7 fixes: playback_finished events, avatar keep-alive, blocked send task
livekit-agents>=1.3.7
livekit-plugins-deepgram>=1.3.0
livekit-plugins-cartesia>=1.3.0
livekit-plugins-bey>=1.3.0
livekit-plugins-silero>=1.3.0
livekit-plugins-turn-detector>=1.3.0
# Noise cancellation for improved turn detection
livekit-plugins-noise-cancellation>=0.2.0

# HTTP client for API calls
httpx>=0.27.0
aiohttp>=3.9.0

# Environment and config
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Convex client
convex>=0.6.0

# Utilities
tenacity>=8.2.0
structlog>=24.0.0
Pillow>=10.0.0

# RAG (Retrieval Augmented Generation)
zep-cloud>=2.0.0

# Web Search
tavily-python>=0.3.0

# OCR for scanned PDFs
paddleocr>=2.9.0
paddlepaddle>=2.6.0
pymupdf>=1.24.0

# Document processing
python-docx>=1.0.0
python-pptx>=0.6.0

# OCR Server
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.9
